(S) BEL-1378 Refactoring tests
test_container was getting too huge. Moved autoscaling testing into
it's own file, and made the fixtures for a container usable with behaves_like
to DRY things up.
